app/requirements.txt
Flask
Flask-SocketIO
Flask-WTF
Flask-SQLAlchemy
flask-session
Flask-Sitemap
SQLAlchemy
Werkzeug
groqcloud
anthropic 
openai 
ragie
pytz
regex
requests
langchain
langchain-community
langchain-openai 
langchain-groq 
langchain-anthropic
langchain-core
langchain-google-genai
langchain-ragie
Werkzeug
PyJWT
gunicorn
gevent
gevent-websocket
eventlet
markdown
markdown-urlize
beautifulsoup4
ragflow-sdk
